changes for folder migration
