+++ server-utils/server_utils/persistence_directory.py
@@ -1,0 +1,93 @@
+"""Create or reset a server's persistence directory.
+
+Shared utilities used by both robot-server and auth-server.
+"""
+
+from logging import getLogger
+from pathlib import Path
+from shutil import rmtree
+from tempfile import mkdtemp
+from typing import Optional
+
+from anyio import Path as AsyncPath
+from anyio import to_thread
+from typing_extensions import Final
+
+from server_utils.folder_migrator import MigrationOrchestrator
+
+_RESET_MARKER_FILE_NAME: Final = "_TO_BE_DELETED_ON_REBOOT"
+_RESET_MARKER_FILE_CONTENTS: Final = """\
+This file was placed here by an Opentrons server.
+It tells the server to clear this directory on the next boot,
+after which it will delete this file.
+"""
+
+_log = getLogger(__name__)
+
+
+class PersistenceResetter:
+    """Mark a persistence directory for deletion on the next boot.
+
+    Deletions are deferred to the next boot so that in-progress HTTP requests
+    and background tasks are not disrupted.
+    """
+
+    def __init__(self, directory_to_reset: Path) -> None:
+        self._directory_to_reset = directory_to_reset
+
+    async def mark_directory_reset(self) -> None:
+        """Write a marker file that ``prepare_root`` will honour on the next boot."""
+        file = AsyncPath(self._directory_to_reset / _RESET_MARKER_FILE_NAME)
+        await file.write_text(encoding="utf-8", data=_RESET_MARKER_FILE_CONTENTS)
+
+
+async def prepare_root(
+    persistence_directory_root: Optional[Path],
+    temp_dir_prefix: str,
+) -> Path:
+    """Return ``persistence_directory_root`` after preparing it, if necessary.
+
+    This will create the directory if it doesn't already exist,
+    and clear its contents if it was previously marked for reset.
+
+    If ``persistence_directory_root`` is ``None``, this will return a fresh
+    temporary directory whose name starts with *temp_dir_prefix*.
+    """
+    if persistence_directory_root is None:
+        new_temporary_directory = Path(mkdtemp(prefix=temp_dir_prefix))
+        _log.info(
+            f"Using auto-created temporary directory {new_temporary_directory}"
+            f" for persistence."
+        )
+        return new_temporary_directory
+
+    if await _is_marked_for_reset(directory_to_reset=persistence_directory_root):
+        _log.info(f"{persistence_directory_root} was marked for reset. Deleting it.")
+        # FIXME(mm, 2024-01-23): This can leave the persistence directory
+        # in a half-deleted state if it deletes the marker file, and then some
+        # of the other files, and then the device is power-cycled before it can
+        # finish.
+        await to_thread.run_sync(rmtree, persistence_directory_root)
+
+    await AsyncPath(persistence_directory_root).mkdir(parents=True, exist_ok=True)
+    _log.info(f"Using directory {persistence_directory_root} for persistence.")
+    return persistence_directory_root
+
+
+async def prepare_active_subdirectory(
+    migration_orchestrator: MigrationOrchestrator,
+) -> Path:
+    """Run pending migrations and return the active version subdirectory.
+
+    This cleans up any stray temporary files left by a prior interrupted
+    migration, then runs all outstanding migrations to bring the database
+    up to the latest version.
+    """
+    await to_thread.run_sync(migration_orchestrator.clean_up_stray_temp_files)
+    subdirectory = await to_thread.run_sync(migration_orchestrator.migrate_to_latest)
+    return subdirectory
+
+
+async def _is_marked_for_reset(directory_to_reset: Path) -> bool:
+    """Return whether the persistence directory has been marked to be reset."""
+    return await (AsyncPath(directory_to_reset) / _RESET_MARKER_FILE_NAME).exists()
